+++ common/src/test/java/com/amazonaws/services/schemaregistry/common/UserAgentRequestInterceptorTest.java
@@ -13,8 +13,6 @@ import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.core.ApiName;
 import software.amazon.awssdk.core.SdkRequest;
 import software.amazon.awssdk.core.interceptor.Context;
-import software.amazon.awssdk.services.glue.GlueClient;
-import software.amazon.awssdk.services.glue.GlueClientBuilder;
 import software.amazon.awssdk.services.glue.model.GetSchemaVersionRequest;
 
 import java.util.stream.Stream;