+++ CHANGELOG.md
@@ -69,5 +69,6 @@ GlueSchemaRegistryKafkaSerializer/GlueSchemaRegistryKafkaDeserializer.
 
 ## Release 2.0.0
 * Add GlueClientBuilder as an argument to AWSSchemaRegistryClient
+* Change default http client implementation to apache http client
 * Remove url-connection-client as a runtime dependency
 * Remove proxyUrl from accessible configuration